fix: split utils into seperate classes
