+++ src/utils/bem-class/bem-class.ts
@@ -1,0 +1,3 @@
+export function bemCls(this: string, element?: string, modifier?: string) {
+	return `${this}${element ? `__${element}` : ''}${modifier ? `--${modifier}` : ''}`;
+}
