+++ src/utils/bem-class/index.ts
@@ -1,3 +1,1 @@
-export function bemCls(this: string, element?: string, modifier?: string) {
-	return `${this}${element ? `__${element}` : ''}${modifier ? `--${modifier}` : ''}`;
-}
+export * from './bem-class';
