+++ src/utils/variant-classes/variant-classes.ts
@@ -1,0 +1,13 @@
+export const getVariantClasses = (
+	baseClass: string,
+	variants?: string | string[],
+	separator = '--'
+): string[] => {
+	if (!baseClass || !variants) {
+		return [];
+	}
+
+	const variantsArray = Array.isArray(variants) ? variants : [variants];
+
+	return variantsArray.map((variant) => `${baseClass}${separator}${variant}`);
+};
